Add Add a New Sheet, Insert Row methods + Refactoring
