+++ payload-functions.js
@@ -1,3 +1,5 @@
+const { extractSpreadsheetIdFromUrl } = require("./helpers");
+
 function prepareStartSpreadsheetPayload(params) {
   return {
     requestBody: {
@@ -15,6 +17,36 @@ function prepareStartSpreadsheetPayload(params) {
   };
 }
 
+function prepareAddSheetPayload(params) {
+  return {
+    spreadsheetId: extractSpreadsheetIdFromUrl(params.spreadsheetUrl),
+    requestBody: {
+      requests: [{
+        addSheet: {
+          properties: {
+            title: params.sheetTitle,
+          },
+        },
+      }],
+    },
+  };
+}
+
+function prepareInsertRowPayload(params) {
+  const range = `${params.sheetTitle}!A${params.row}`;
+  return {
+    spreadsheetId: extractSpreadsheetIdFromUrl(params.spreadsheetUrl),
+    range,
+    valueInputOption: "USER_ENTERED",
+    requestBody: {
+      majorDimension: "ROWS",
+      values: [params.data],
+    },
+  };
+}
+
 module.exports = {
   prepareStartSpreadsheetPayload,
+  prepareAddSheetPayload,
+  prepareInsertRowPayload,
 };
